+++ crates/pallets/starknet/src/lib.rs
@@ -93,7 +93,7 @@ use sp_runtime::DigestItem;
 use starknet_api::api_core::{ChainId, CompiledClassHash, ContractAddress, EntryPointSelector, Nonce};
 use starknet_api::block::{BlockNumber, BlockTimestamp};
 use starknet_api::deprecated_contract_class::EntryPointType;
-use starknet_api::hash::StarkFelt;
+use starknet_api::hash::{StarkFelt, StarkHash};
 use starknet_api::state::StorageKey;
 use starknet_api::transaction::{Calldata, Event as StarknetEvent, Fee, MessageToL1, TransactionHash};
 use starknet_crypto::FieldElement;
@@ -108,6 +108,8 @@ pub(crate) const LOG_TARGET: &str = "runtime::starknet";
 
 pub const ETHEREUM_EXECUTION_RPC: &[u8] = b"starknet::ETHEREUM_EXECUTION_RPC";
 pub const ETHEREUM_CONSENSUS_RPC: &[u8] = b"starknet::ETHEREUM_CONSENSUS_RPC";
+
+pub const SN_OS_CONFIG_HASH_VERSION: &str = "StarknetOsConfig1";
 
 // syntactic sugar for logging.
 #[macro_export]
@@ -1148,6 +1150,15 @@ impl<T: Config> Pallet<T> {
 
     pub fn program_hash() -> Felt252Wrapper {
         T::ProgramHash::get()
+    }
+
+    pub fn config_hash() -> StarkHash {
+        T::SystemHash::compute_hash_on_elements(&[
+            FieldElement::from_byte_slice_be(SN_OS_CONFIG_HASH_VERSION.as_bytes()).unwrap(),
+            T::ChainId::get().into(),
+            Self::fee_token_address().0.0.into(),
+        ])
+        .into()
     }
 
     pub fn is_transaction_fee_disabled() -> bool {